+++ src/aosm/azext_aosm/configuration_models/onboarding_vnf_input_config.py
@@ -2,10 +2,13 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
+from __future__ import annotations
+from azure.cli.core.azclierror import ValidationError
 from dataclasses import dataclass, field
 from azext_aosm.configuration_models.onboarding_nfd_base_input_config import OnboardingNFDBaseInputConfig
 from azext_aosm.configuration_models.common_input import ArmTemplatePropertiesConfig
+
+from typing import List
 
 
 @dataclass
@@ -24,7 +27,7 @@ class VhdImageConfig:
         metadata={
             "comment": (
                 "Optional. File path of the artifact you wish to upload from your local disk. "
-                "Delete if not required. Relative paths are relative to the configuration file. "
+                "Delete if not required.\nRelative paths are relative to the configuration file. "
                 "On Windows escape any backslash with another backslash."
             )
         }
@@ -33,51 +36,64 @@ class VhdImageConfig:
         default="",
         metadata={
             "comment": (
-                "Optional. SAS URL of the blob artifact you wish to copy to your Artifact Store. "
+                "Optional. SAS URL of the blob artifact you wish to copy to your Artifact Store.\n"
                 "Delete if not required. "
                 "On Windows escape any backslash with another backslash."
             )
         }
     )
-    image_disk_size_GB: str = field(
+    image_disk_size_GB: str | None = field(
         default="",
         metadata={
             "comment": (
-                "Optional. Specifies the size of empty data disks in gigabytes. "
+                "Optional. Specifies the size of empty data disks in gigabytes.\n"
                 "This value cannot be larger than 1023 GB. Delete if not required."
             )
         }
     )
-    image_hyper_v_generation: str = field(
+    image_hyper_v_generation: str | None = field(
         default="",
         metadata={
             "comment": (
-                "Optional. Specifies the HyperVGenerationType of the VirtualMachine created from the image. "
+                "Optional. Specifies the HyperVGenerationType of the VirtualMachine created from the image.\n"
                 "Valid values are V1 and V2. V1 is the default if not specified. Delete if not required."
             )
         }
     )
-    image_api_version: str = field(
+    image_api_version: str | None = field(
         default="",
         metadata={
             "comment": (
-                "Optional. The ARM API version used to create the Microsoft.Compute/images resource. "
+                "Optional. The ARM API version used to create the Microsoft.Compute/images resource.\n"
                 "Delete if not required."
             )
         }
     )
+    def validate(self):
+        """Validate the configuration."""
+        if not self.version:
+            raise ValidationError("Artifact version must be set")
+        if "." not in self.version or "-" in self.version:
+            raise ValidationError(
+                "Config validation error. VHD image artifact version should be in"
+                " format A.B.C"
+            )
+        if self.blob_sas_url and self.file_path:
+            raise ValidationError("Only one of file_path or blob_sas_url may be set for vhd.")
+        if not (self.blob_sas_url or self.file_path):
+            raise ValidationError("One of file_path or sas_blob_url must be set for vhd.")
 
 
 @dataclass
 class OnboardingVNFInputConfig(OnboardingNFDBaseInputConfig):
     """Input configuration for onboarding VNFs."""
 
-    blob_artifact_store_name: str = field(
+    blob_artifact_store_name: str | None = field(
         default="",
         metadata={
             "comment": (
-                "Optional. Name of the storage account Artifact Store resource. "
-                " Will be created if it does not exist (with a default name if none is supplied)."
+                "Optional. Name of the storage account Artifact Store resource. \n"
+                "Will be created if it does not exist (with a default name if none is supplied)."
             )
         }
     )
@@ -92,11 +108,45 @@ class OnboardingVNFInputConfig(OnboardingNFDBaseInputConfig):
     )
 
     # TODO: Add better comments
-    arm_template: [ArmTemplatePropertiesConfig] = field(
+    arm_template: List[ArmTemplatePropertiesConfig] = field(
         default_factory=lambda: [ArmTemplatePropertiesConfig()],
         metadata={"comment": "ARM template configuration."},
     )
 
-    vhd: [VhdImageConfig] = field(
+    vhd: List[VhdImageConfig] = field(
         default_factory=lambda: [VhdImageConfig()],
         metadata={"comment": "VHD image configuration."})
+
+    def __post_init__(self):
+        for arm_template in self.arm_template:
+            arm_list = []
+            if arm_template and isinstance(arm_template, dict):
+                arm_list.append(ArmTemplatePropertiesConfig(**arm_template))
+            else:
+                arm_list.append(arm_template)
+        self.arm_template = arm_list
+
+        for vhd in self.vhd:
+            vhd_list = []
+            if vhd and isinstance(vhd, dict):
+                vhd_list.append(VhdImageConfig(**vhd))
+            else:
+                vhd_list.append(vhd)
+        self.vhd = vhd_list
+
+    def validate(self):
+        """Validate the configuration."""
+        super().validate()
+
+        if not self.image_name_parameter:
+            raise ValidationError("image_name_parameter must be set")
+        if not self.arm_template:
+            raise ValidationError("arm_template must be set")
+        if not self.vhd:
+            raise ValidationError("vhd must be set")
+        if not self.arm_template:
+            raise ValidationError("You must include at least one arm template")
+        for arm_template in self.arm_template:
+            arm_template.validate()
+        for vhd in self.vhd:
+            vhd.validate()