+++ src/aosm/azext_aosm/definition_folder/builder/artifact_builder.py
@@ -10,11 +10,13 @@ from .base_builder import BaseDefinitionElementBuilder
 
 from azext_aosm.common.artifact import BaseArtifact
 
+from typing import List
+
 class ArtifactDefinitionElementBuilder(BaseDefinitionElementBuilder):
     """ Artifact builder """
-    artifacts: [BaseArtifact]
+    artifacts: List[BaseArtifact]
 
-    def __init__(self, path: Path, artifacts: [BaseArtifact], only_delete_on_clean: bool = False):
+    def __init__(self, path: Path, artifacts: List[BaseArtifact], only_delete_on_clean: bool = False):
         super().__init__(path, only_delete_on_clean)
         self.artifacts = artifacts